commands: Create compilePipeline function
